Reset Password Implemented
Functionality to reset user password is Implemented
A reset url will be send to the user email for that
For localhost testing a local SMPT server is created(smtp_server.py)
which simulates email send and receive.
Actual email is not being sent to the recepient
For production add the required smpt details.

For checking reset password locally
1. Run smtp_server.py script in another terminal
2 Go to reset passord page
3. Enter the email to reset password for
4. Check the console of locally run smtp server
5. Use the link printed in console to change password.